backend/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
pgvector==0.3.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# OpenAI
openai==1.51.0
tiktoken==0.7.0

# Document processing
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==10.4.0
python-docx==1.1.2

# Excel reports
openpyxl==3.1.5

# Telegram
python-telegram-bot==21.6

# AWS S3 (async)
aioboto3==13.1.1

# Utils
pydantic-settings==2.5.2
aiofiles==24.1.0
python-dateutil==2.9.0
